refactor dag to use gusty, split tasks into separate files
