+++ .pre-commit-config.yaml
@@ -8,6 +8,7 @@ repos:
       - id: trailing-whitespace
       - id: end-of-file-fixer
       - id: check-yaml
+        args: ["--unsafe"]
       - id: check-added-large-files
   - repo: https://github.com/psf/black
     rev: 19.10b0
