+++ airflow/requirements.txt
@@ -2,3 +2,4 @@ gcsfs==0.7.2
 intake==0.6.1
 ipdb==0.13.4
 pandas==1.1.4
+gusty==0.5.1
